refactor(send): execute prepared transaction in saga
